requirements.txt
transformers==4.46.3
PyPDF2==3.0.1
sentence_transformers==3.3.1
accelerate==1.1.1
streamlit==1.40.1
python-docx==1.1.2
python-pptx==1.0.2
